Add topology requirement sourcing from ModelSpec & enforcement
