+++ workflows/workflow_types.py
@@ -2,7 +2,7 @@
 #
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
-from enum import IntEnum, auto
+from enum import Enum, IntEnum, auto
 
 
 class WorkflowType(IntEnum):
@@ -94,6 +94,19 @@ class DeviceTypes(IntEnum):
             raise ValueError(f"Invalid DeviceType: {self}")
         return mapping[self]
 
+    def get_topology_requirement(self) -> bool:
+        """Return the required system-level mesh topology for a given DeviceType"""
+        # topology not required for Blackhole
+        if self.is_blackhole():
+            return
+
+        # mesh topology only required for multi-wh configurations, excluding galaxy
+        requires_mesh_topology = {DeviceTypes.N150X4, DeviceTypes.T3K}
+        if self in requires_mesh_topology:
+            return SystemTopology.MESH
+
+        # TODO: for future, more advanced topology requirements
+
     def is_wormhole(self) -> bool:
         wormhole_devices = {
             DeviceTypes.N150,
@@ -128,6 +141,21 @@ class DeviceTypes(IntEnum):
                 f"Invalid DeviceType or data_parallel: {self}, {data_parallel}"
             )
         return data_parallel_map[(self, data_parallel)]
+
+
+class SystemTopology(Enum):
+    """Enumerates all valid Wormhole system topologies"""
+    MESH = "Mesh"
+    LINEAR_TORUS = "Linear/Torus"
+    ISOLATED = "Isolated or not configured"
+
+    @classmethod
+    def from_topology_string(cls, value: str):
+        """Instantiates a SystemTopology from the result string from the `tt-topology -ls` command"""
+        for member in cls:
+            if member.value.lower() == value.lower():  # case-insensitive match
+                return member
+        raise ValueError(f"Unknown topology configuration: {value}")
 
 
 class ReportCheckTypes(IntEnum):
